Finished basic building script for QPrompt's Qt 6 build

+++ setup.sh
@@ -22,7 +22,12 @@
 #**************************************************************************
 
 cat << EOF
-usage: $0 [-G <[DEB|AppImage|EXE|DMG]>| -n <levels>][--help][--version]
+usage: $0 <CMAKE_BUILD_TYPE> <CMAKE_PREFIX_PATH> [<QT_MAJOR_VERSION> | ] [CLEAR | CLEAR_ALL]
+
+Defaults:
+ * CMAKE_BUILD_TYPE: "Release"
+ * CMAKE_PREFIX_PATH: "~/Qt/6.6.1/gcc_64/"
+ * QT_MAJOR_VERSION: "6"
 
 Setup script for building QPrompt
 This script assumes you've already installed the following dependencies:
@@ -31,19 +36,54 @@ This script assumes you've already installed the following dependencies:
  * Qt 6 Open Source
 EOF
 
-CMAKE_BUILD_TYPE="Release"
-CMAKE_PREFIX_PATH="~/Qt/6.7.0/gcc_64/"
-CMAKE_INSTALL_PREFIX="./install"
+CMAKE_BUILD_TYPE=$1
+if [ "$CMAKE_BUILD_TYPE" == "" ]
+        then CMAKE_BUILD_TYPE="Release"
+fi
+CMAKE_PREFIX_PATH=$2
+if [ "$CMAKE_PREFIX_PATH" == "" ]
+        then CMAKE_PREFIX_PATH="~/Qt/6.6.1/gcc_64/"
+fi
+QT_MAJOR_VERSION=$3
+if [ "$QT_MAJOR_VERSION" == "" ]
+        then QT_MAJOR_VERSION="5"
+fi
+CLEAR_ARG="${@: -1}"
+if [ "$CLEAR_ARG" == "CLEAR" ]
+    then
+    CLEAR=true
+    CLEAR_ALL=false
+elif [ "$CLEAR_ARG" == "CLEAR_ALL" ]
+    then
+    CLEAR=true
+    CLEAR_ALL=true
+else
+    CLEAR=false
+    CLEAR_ALL=false
+fi
+
+# Constants
+CMAKE_INSTALL_PREFIX="./build/dist"
+ARCHITECTURE="$(uname -m)"
 
 echo "Build directory is ./build"
-rm -dRf ./build ./install
+if $CLEAR_ALL # QPrompt and dependencies
+    then
+    rm -dRf ./build ./install
+elif $CLEAR # QPrompt
+    then
+    rm -dRf ./build
+fi
 mkdir -p build install
 
 echo "Downloading git submodules"
 git submodule update --init --recursive
 
 echo "Extra CMake Modules"
-rm -dRf ./3rdparty/extra-cmake-modules/build
+if $CLEAR_ALL
+    then
+    rm -dRf ./3rdparty/extra-cmake-modules/build
+fi
 cmake -DCMAKE_BUILD_TYPE=$CMAKE_BUILD_TYPE -DCMAKE_PREFIX_PATH=$CMAKE_PREFIX_PATH -DCMAKE_INSTALL_PREFIX=$CMAKE_INSTALL_PREFIX -B ./3rdparty/extra-cmake-modules/build ./3rdparty/extra-cmake-modules/
 cmake --build ./3rdparty/extra-cmake-modules/build
 cmake --install ./3rdparty/extra-cmake-modules/build
@@ -51,7 +91,10 @@ cmake --install ./3rdparty/extra-cmake-modules/build
 # KDE Frameworks
 for dependency in ./3rdparty/k*; do
     echo $dependency
-    rm -dRf ./$dependency/build
+    if $CLEAR_ALL
+    then
+        rm -dRf ./$dependency/build
+    fi
     cmake -DCMAKE_BUILD_TYPE=$CMAKE_BUILD_TYPE -DCMAKE_PREFIX_PATH=$CMAKE_PREFIX_PATH -DCMAKE_INSTALL_PREFIX=$CMAKE_INSTALL_PREFIX -B ./$dependency/build ./$dependency/
     cmake --build ./$dependency/build
     cmake --install ./$dependency/build
